Script for partial automation of variant analysis
This was created specifically to run on a 20 core machine - therefore some commands will need to be changed. It also assumes that all programs are going to already be in your PATH. This script is a VERY basic one, and was simply forced together to make the process slightly more automated for people less confident using command line.