Partial automation of variant analysis
This is a very basic script forced together to help out people not as confident in command line. The script is optimised to be used on a 20 core machine (so changes may need to be made). It was also created assuming that all programs used are currently in your PATH, this was intentionally done for ease of use in house.